# Conflicts:
#	src/main/java/com/lms/cmpe/controller/BookController.java
#	src/main/java/com/lms/cmpe/model/WaitlistBooksToBeAssigned.java
#	src/main/resources/templates/profile.html

+++ src/main/java/com/lms/cmpe/controller/BookController.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
 import java.util.ArrayList;
@@ -38,6 +39,11 @@ public class BookController {
 
     @GetMapping("/books")
     public String getAllBooks(Model model, HttpSession session){
+
+        if(session.getAttribute("user")==null){
+            return "redirect:/";
+        }
+
             model.addAttribute("user",session.getAttribute("user"));
             model.addAttribute("books",bookService.getAllBooks());
 
@@ -88,15 +94,26 @@ public class BookController {
     }
 
     @GetMapping("/book")
-    public String bookForm(Model model,HttpSession session){
+    public String bookForm(Model model, HttpSession session){
+
+        if(session.getAttribute("user")==null){
+            return "redirect:/";
+        }
+
         Book book = new Book();
         model.addAttribute("user",session.getAttribute("user"));
         model.addAttribute("book",book);
+
         return "book";
     }
 
     @RequestMapping(value = "/book/add", method = RequestMethod.POST)
-    public String addBook(@ModelAttribute Book book, Model model, @RequestParam(value="action") String action, HttpSession session){
+    public String addBook(@ModelAttribute Book book, Model model, @RequestParam(value="action") String action,
+                          HttpSession session, RedirectAttributes redirectAttributes){
+        if(session.getAttribute("user")==null){
+            return "redirect:/";
+        }
+
         model.addAttribute("user",session.getAttribute("user"));
         if(action.equals("addkeyword")){
             book.getBookKeywordsList().add(new BookKeywords());
@@ -110,7 +127,8 @@ public class BookController {
             int total = book.getNoOfAvailableCopies();
             book.setNoOfAvailableCopies(total);
             bookService.addBook(book);
-            return "redirect:/book";
+            redirectAttributes.addFlashAttribute("message", "Successfully Added! Add more books..");
+            return"redirect:/book";
         }
         return "test";
     }
@@ -130,6 +148,10 @@ public class BookController {
 
     @GetMapping("/book/update/{id}")
     public String bookUpdateForm(@PathVariable("id") int id, Model model, HttpSession session){
+        if(session.getAttribute("user")==null){
+            return "redirect:/";
+        }
+
         model.addAttribute("book", bookService.getBookById(id));
         model.addAttribute("user", session.getAttribute("user"));
         return "updatebook";
@@ -138,7 +160,11 @@ public class BookController {
     @RequestMapping(value = "/book/update/{id}",method = RequestMethod.POST)
     public String updateBook(@ModelAttribute Book book, @PathVariable("id") int id,
                              @RequestParam(value="action", required=true) String action, HttpSession session,
-                             Model model){
+                             Model model, RedirectAttributes redirectAttributes){
+        if(session.getAttribute("user")==null){
+            return "redirect:/";
+        }
+
         model.addAttribute("user",session.getAttribute("user"));
         book.setBookId(id);
 
@@ -152,21 +178,31 @@ public class BookController {
         if(action.equals("update")){
             book.addBookKeywords();
             bookService.updateBook(book);
+            redirectAttributes.addFlashAttribute("message","Book info is updated");
         }
 
         return String.format("redirect:/book/update/%d",book.getBookId());
     }
 
     @GetMapping("/book/delete/{id}")
-    public String deleteBookByid(@PathVariable("id") int id){
+    public String deleteBookByid(@PathVariable("id") int id, HttpSession session, RedirectAttributes redirectAttributes){
+        if(session.getAttribute("user")==null){
+            return "redirect:/";
+        }
+
         Book book= bookService.getBookById(id);
         bookService.deleteBook(book);
+        redirectAttributes.addFlashAttribute("message", "Book has been deleted");
         return "redirect:/books";
     }
 
     @GetMapping("/book/addtocart/{id}")
     public String addBookToCart(@PathVariable("id") int id, @ModelAttribute BookList booklist,
                                 Model model, HttpSession session){
+            if(session.getAttribute("user")==null){
+                return "redirect:/";
+            }
+
             booklist = (BookList)session.getAttribute("booklist");
 
             for(Book book:booklist.getBookList()){
@@ -183,6 +219,10 @@ public class BookController {
     public String removeFromCart(@PathVariable("id") int id, @ModelAttribute BookList booklist
                                 ,@PathVariable("index") int index, HttpSession session){
 
+        if(session.getAttribute("user")==null){
+            return "redirect:/";
+        }
+
         booklist = (BookList)session.getAttribute("booklist");
         booklist.getBookList().remove(index);
         return "redirect:/books";
@@ -190,6 +230,10 @@ public class BookController {
 
     @GetMapping("/books/checkout")
     public String checkout(HttpSession session, Model model){
+
+        if(session.getAttribute("user")==null){
+            return "redirect:/";
+        }
 
         BookList bookList= (BookList)session.getAttribute("booklist");
         System.out.println(bookList.toString());
@@ -239,6 +283,10 @@ public class BookController {
 
     @GetMapping("/books/searchresults")
     public String searchResults(Model model, HttpSession session){
+        if(session.getAttribute("user")==null){
+            return "redirect:/";
+        }
+
         model.addAttribute("user", session.getAttribute("user"));
         model.addAttribute("booklist", session.getAttribute("booklist"));
         model.addAttribute("books", session.getAttribute("books"));
@@ -259,6 +307,10 @@ public class BookController {
 
     @PostMapping(value = "/books", params = "search")
     public String search(Model model, HttpSession session, @RequestParam("keyword") String keyword){
+        if(session.getAttribute("user")==null){
+            return "redirect:/";
+        }
+
         model.addAttribute("user", session.getAttribute("user"));
         model.addAttribute("books", bookService.getBooksByKey(keyword));
         session.setAttribute("books",bookService.getBooksByKey(keyword));
@@ -269,6 +321,10 @@ public class BookController {
 
     @GetMapping("/book/delete/searchresult/{id}")
     public String deleteBookSearchedByid(@PathVariable("id") int id, HttpSession session){
+        if(session.getAttribute("user")==null){
+            return "redirect:/";
+        }
+
         Book book= bookService.getBookById(id);
         bookService.deleteBook(book);
         ArrayList<Book> templist = (ArrayList<Book>)session.getAttribute("books");
@@ -285,6 +341,10 @@ public class BookController {
     @GetMapping("/book/addtocart/searchresult/{id}")
     public String addBookSearchedToCart(@PathVariable("id") int id, @ModelAttribute BookList booklist,
                                 Model model, HttpSession session){
+        if(session.getAttribute("user")==null){
+            return "redirect:/";
+        }
+
         booklist = (BookList)session.getAttribute("booklist");
 
         for(Book book:booklist.getBookList()){
@@ -300,9 +360,24 @@ public class BookController {
     public String removeSearchResultFromCart(@PathVariable("id") int id, @ModelAttribute BookList booklist
             ,@PathVariable("index") int index, HttpSession session){
 
+        if(session.getAttribute("user")==null){
+            return "redirect:/";
+        }
+
         booklist = (BookList)session.getAttribute("booklist");
         booklist.getBookList().remove(index);
         return "redirect:/books/searchresults";
     }
+
+    @GetMapping("reissue/book/{id}")
+    public String reIssueBook(@PathVariable("id") int id, HttpSession session)
+    {
+        System.out.println("Reissue Hereeeeeeeeeeeeeeeee " + id);
+        User user = (User)session.getAttribute("user");
+        transactionService.reissueBook(id, user.getUserId());
+        return "redirect:/profile";
+
+    }
+
 
 }
